Tweaked Tile.show to be injected only when rf_ipython is imported and running in ipython environment.
